requirements.txt
fastapi
uvicorn
httpx
pytest
pytest-asyncio
pytest-cov
python-dotenv
